fixup! fixup! feat(2d): add Polygon component (#455)
