+++ packages/2d/src/components/Polygon.ts
@@ -1,33 +1,8 @@
-import {
-  // PossibleSpacing,
-  BBox,
-  // SpacingSignal,
-} from '@motion-canvas/core/lib/types';
+import {BBox} from '@motion-canvas/core/lib/types';
 import {Shape, ShapeProps} from './Shape';
 import {initial, signal} from '../decorators';
 import {SignalValue, SimpleSignal} from '@motion-canvas/core/lib/signals';
-
-function drawPolygon(
-  path: CanvasRenderingContext2D | Path2D,
-  rect: BBox,
-  // width: number,
-  // height: number,
-  sides: number,
-) {
-  const width = rect.width / 2;
-  const height = rect.height / 2;
-  for (let i = 0; i <= sides; i += 1) {
-    const theta = (i * 2 * Math.PI) / sides;
-    const x = width * Math.sin(theta);
-    const y = -height * Math.cos(theta);
-    if (i == 0) {
-      path.moveTo(x, y);
-    } else {
-      path.lineTo(x, y);
-    }
-  }
-  path.closePath();
-}
+import {drawPolygon} from '../utils';
 
 export interface PolygonProps extends ShapeProps {
   /**
@@ -36,30 +11,54 @@ export interface PolygonProps extends ShapeProps {
   sides?: SignalValue<number>;
 }
 
+/**
+ * A node for drawing regular polygons.
+ *
+ * @remarks
+ * This node can be used to render shapes such as: triangle, pentagon,
+ * hexagon and more.
+ *
+ * Note that the polygon is inscribed in a circle defined by the height
+ * and width. If height and width are unequal, the polygon is inscribed
+ * in the resulting ellipse.
+ *
+ * Since the polygon is inscribed in the circle, the actual displayed
+ * height and width may differ somewhat from the bounding rectangle. This
+ * will be particularly noticable if the number of sides is low, e.g. for a
+ * triangle.
+ *
+ * @example
+ * A hexagon:
+ * ```tsx
+ * <Polygon
+ *   sides={6}
+ *   size={300}
+ *   fill={'lightseagreen'}
+ * />
+ * ```
+ * A pentagon outline:
+ * ```tsx
+ * <Polygon
+ *   sides={5}
+ *   size={300}
+ *   stroke={'lightblue'}
+ *   lineWidth={8}
+ * />
+ * ```
+ */
 export class Polygon extends Shape {
   /**
-   * Sets the number of sides of the polygon. For example, a value of 6
-   * creates a hexagon.
+   * Sets the number of sides of the polygon.
    *
    * @remarks
-   * Note that the polygon is inscribed in a circle defined by the height
-   * and width. If height and width are unequal, the polygon is inscribed
-   * in the resulting ellipse.
-   *
-   * Since the polygon is inscribed in the circle, the actual displayed
-   * height and width may differ somewhat from the bounding rectangle. This
-   * will be particularly noticable if the number of sides is low, e.g. for a
-   * triangle.
+   * For example, a value of 6 creates a hexagon.
    *
    *
    * @example
    * ```tsx
    * <Polygon
-   *   x={0}
-   *   y={0}
-   *   width={320}
-   *   height={320}
-   *   sides={6}
+   *   size={320}
+   *   sides={7}
    *   stroke={'#fff'}
    *   lineWidth={8}
    *   fill={'lightseagreen'}
@@ -88,7 +87,6 @@ export class Polygon extends Shape {
     const sides = this.sides();
     const rippleSize = this.rippleSize();
 
-    // const box = BBox.fromSizeCentered(this.size());
     const box = BBox.fromSizeCentered(this.size()).expand(rippleSize);
     drawPolygon(path, box, sides);
