+++ packages/2d/src/utils/CanvasUtils.ts
@@ -162,6 +162,25 @@ export function strokeRect(context: CanvasRenderingContext2D, rect: BBox) {
   context.strokeRect(rect.x, rect.y, rect.width, rect.height);
 }
 
+export function drawPolygon(
+  path: CanvasRenderingContext2D | Path2D,
+  rect: BBox,
+  sides: number,
+) {
+  const size = rect.size.scale(0.5);
+  for (let i = 0; i <= sides; i++) {
+    const theta = (i * 2 * Math.PI) / sides;
+    const direction = Vector2.fromRadians(theta).perpendicular;
+    const vertex = direction.mul(size);
+    if (i == 0) {
+      path.moveTo(vertex.x, vertex.y);
+    } else {
+      path.lineTo(vertex.x, vertex.y);
+    }
+  }
+  path.closePath();
+}
+
 export function drawImage(
   context: CanvasRenderingContext2D,
   image: CanvasImageSource,
